\name{data1}
\alias{data1}
\docType{data}
\title{A toy dataset}
\description{
  The \code{data1} dataset has 100 rows and 4 columns.
}
\usage{data(data1)}
\format{
  A data frame with 100 observations on the following 4 variables.
  \describe{
    \item{\code{y}}{a gamma-distributed response variable}
    \item{\code{fat1}}{a four-level factor}
    \item{\code{x1}}{a numeric covariate}
    \item{\code{x2}}{a numeric covariate}

  }
}
\details{
  This is a toy dataset used to show how function \code{shglm} works.
}

\examples{
data(data1)
}
\keyword{datasets}
