\name{speedglm-package}
\alias{speedglm-package}
\docType{package}
\title{
Fitting Linear and Generalized Linear Models to Large Data Sets
}
\description{
Fits Linear and Generalized Linear Models to large data sets. For data loaded in R memory the fitting is usually
fast, especially if R is linked against an optimized BLAS. For data sets of size greater 
of R memory, the fitting is made by an updating algorithm.}
\details{
\tabular{ll}{
Package: \tab speedglm\cr
Type: \tab Package\cr
Version: \tab 0.3-4\cr
Date: \tab 2022-02-19\cr
Depends: \tab Matrix, stats, MASS\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Enea <marco.enea@unipa.it>, with contributions from Ronen Meiri and Tomer Kalimi (on behalf of DMWay Analytics LTD).

Maintainer: Marco Enea <marco.enea@unipa.it> 

}

\keyword{ models}


       