\name{speedglm-package}
\alias{speedglm-package}
\docType{package}
\title{
Fitting Linear and Generalized Linear Models to Large Data Sets.
}
\description{
Fits LMs and GLMs to large data sets. For data loaded in R memory the fitting is usually
fast, especially if R is linked against an optimized BLAS. For data sets of size greater 
of R memory, the fitting is made by an updating algorithm.}
\details{
\tabular{ll}{
Package: \tab speedglm\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-07-23\cr
Depends: \tab Matrix\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Enea  

Maintainer: Marco Enea <marco.enea@unipa.it> 

}

\keyword{ models}


       